Refactor new absence application form to make absence date range validation possible
